fix detection of vendored packages
When doing recursive package processing, pass the path to the package
as a root for packages search. This fixes detection of vendored packages
if the vendor directory is somewhere in subdirectory. Before this commit
it finds vendored packages only if they are in "./vendor", where "." is
the current directory of godepgraph process.

Example repo for testing: https://github.com/starius/godepgraph-error